move source code to the src folder
