+++ src/convert.ts
@@ -1,4 +1,4 @@
-const issues = require("./data/source.json") as JiraTicket[];
+const issues = require("../data/source.json") as JiraTicket[];
 import { createWriteStream } from "fs";
 import { exit } from "process";
 import { JiraTicketHistory, Transition, Ticket, JiraTicket, StatusSwitches, StatusSwitchString } from "./types";
